% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{auth}
\alias{auth}
\title{Set AWS credentials for Bioconductor Hubs Ingest endpoint}
\usage{
auth(username, password, endpoint = NULL)
}
\arguments{
\item{username}{Character string of the username provided by administrator}

\item{password}{Character string of the password/key provided by administrator}

\item{endpoint}{Optional custom endpoint URL. If NULL, constructs from username}
}
\value{
Invisible NULL, sets environment variables as side effect
}
\description{
Set AWS credentials for Bioconductor Hubs Ingest endpoint
}
\examples{
\dontrun{
BiocHubsIngestR::auth("myusername", "mypassword")
}
}
