% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{upload}
\alias{upload}
\title{Upload local directory to Bioconductor Hubs Ingest S3 endpoint}
\usage{
upload(path, bucket = "userdata")
}
\arguments{
\item{path}{Character string path to local directory or file to upload}

\item{bucket}{Character string of bucket name, defaults to "userdata"}
}
\value{
Invisible list of uploaded files
}
\description{
Upload local directory to Bioconductor Hubs Ingest S3 endpoint
}
\examples{
\dontrun{
BiocHubsIngestR::upload("path/to/data")
}
}
